# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MediaTracker",
    "version": "0.2.11",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "ApiHeader": []
    }
  ],
  "paths": {
...
    "/api/configuration": {
      "patch": {
        "tags": [
          "Configuration"
        ],
        "operationId": "update",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enableRegistration": {
                    "type": "boolean",
                    "nullable": true
                  },
                  "tmdbLang": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/TmdbLang"
                      }
                    ],
                    "nullable": true
                  },
                  "audibleLang": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/AudibleCountryCode"
                      }
                    ],
                    "nullable": true
                  },
                  "serverLang": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ServerLang"
                      }
                    ],
                    "nullable": true
                  },
                  "igdbClientId": {
                    "type": "string",
                    "nullable": true
                  },
                  "igdbClientSecret": {
                    "type": "string",
                    "nullable": true
                  },
                  "sonarrUrl": {
                    "type": "string",
                    "nullable": true
                  },
                  "sonarrApiKey": {
                    "type": "string",
                    "nullable": true
                  },
                  "watchProvidersCountry": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Configuration"
        ],
        "operationId": "get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "type": "object",
                      "properties": {
                        "enableRegistration": {
                          "type": "boolean"
                        },
                        "tmdbLang": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/TmdbLang"
                            }
                          ],
                          "nullable": true
                        },
                        "audibleLang": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/AudibleCountryCode"
                            }
                          ],
                          "nullable": true
                        },
                        "serverLang": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/ServerLang"
                            }
                          ],
                          "nullable": true
                        },
                        "igdbClientId": {
                          "type": "string",
                          "nullable": true
                        },
                        "igdbClientSecret": {
                          "type": "string",
                          "nullable": true
                        },
                        "sonarrUrl": {
                          "type": "string",
                          "nullable": true
                        },
                        "sonarrApiKey": {
                          "type": "string",
                          "nullable": true
                        },
                        "watchProvidersCountry": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "enableRegistration"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "noUsers": {
                          "type": "boolean"
                        },
                        "demo": {
                          "type": "boolean"
                        },
                        "version": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "demo",
                        "noUsers",
                        "version"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/details/update-metadata/{mediaItemId}": {
      "get": {
        "tags": [
          "MediaItem"
        ],
        "operationId": "updateMetadata",
        "parameters": [
          {
            "name": "mediaItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "password",
                  "username"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                },
                "required": [
                  "confirmPassword",
                  "password",
                  "username"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/UserResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "ApiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "TmdbLang": {
        "enum": [
          "aa",
          "ab",
          "af",
          "am",
          "ar",
          "as",
          "ay",
          "az",
          "ba",
          "be",
          "bg",
          "bh",
          "bi",
          "bn",
          "bo",
          "br",
          "ca",
          "co",
          "cs",
          "cy",
          "da",
          "de",
          "dz",
          "el",
          "en",
          "eo",
          "es",
          "et",
          "eu",
          "fa",
          "fi",
          "fj",
          "fo",
          "fr",
          "fy",
          "ga",
          "gd",
          "gl",
          "gn",
          "gu",
          "ha",
          "he",
          "hi",
          "hr",
          "hu",
          "hy",
          "ia",
          "id",
          "ie",
          "ik",
          "is",
          "it",
          "iu",
          "ja",
          "jw",
          "ka",
          "kk",
          "kl",
          "km",
          "kn",
          "ko",
          "ks",
          "ku",
          "ky",
          "la",
          "ln",
          "lo",
          "lt",
          "lv",
          "mg",
          "mi",
          "mk",
          "ml",
          "mn",
          "mo",
          "mr",
          "ms",
          "mt",
          "my",
          "na",
          "ne",
          "nl",
          "no",
          "oc",
          "om",
          "or",
          "pa",
          "pl",
          "ps",
          "pt",
          "qu",
          "rm",
          "rn",
          "ro",
          "ru",
          "rw",
          "sa",
          "sd",
          "sg",
          "sh",
          "si",
          "sk",
          "sl",
          "sm",
          "sn",
          "so",
          "sq",
          "sr",
          "ss",
          "st",
          "su",
          "sv",
          "sw",
          "ta",
          "te",
          "tg",
          "th",
          "ti",
          "tk",
          "tl",
          "tn",
          "to",
          "tr",
          "ts",
          "tt",
          "tw",
          "ug",
          "uk",
          "ur",
          "uz",
          "vi",
          "vo",
          "wo",
          "xh",
          "yi",
          "yo",
          "za",
          "zh",
          "zu"
        ],
        "type": "string"
      },
      "AudibleCountryCode": {
        "enum": [
          "au",
          "ca",
          "de",
          "es",
          "fr",
          "in",
          "it",
          "jp",
          "uk",
          "us"
        ],
        "type": "string"
      },
      "ServerLang": {
        "enum": [
          "da",
          "de",
          "en",
          "es",
          "fr",
          "ko",
          "pt"
        ],
        "type": "string"
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "admin": {
            "type": "boolean",
            "nullable": true
          },
          "publicReviews": {
            "type": "boolean",
            "nullable": true
          },
          "sendNotificationWhenStatusChanges": {
            "type": "boolean",
            "nullable": true
          },
          "sendNotificationWhenReleaseDateChanges": {
            "type": "boolean",
            "nullable": true
          },
          "sendNotificationWhenNumberOfSeasonsChanges": {
            "type": "boolean",
            "nullable": true
          },
          "sendNotificationForReleases": {
            "type": "boolean",
            "nullable": true
          },
          "sendNotificationForEpisodesReleases": {
            "type": "boolean",
            "nullable": true
          },
          "notificationPlatform": {
            "enum": [
              "Discord",
              "Pushbullet",
              "Pushover",
              "Pushsafer",
              "gotify",
              "ntfy",
              null
            ],
            "type": "string",
            "nullable": true
          },
          "hideOverviewForUnseenSeasons": {
            "type": "boolean",
            "nullable": true
          },
          "hideEpisodeTitleForUnseenEpisodes": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "RequestError": {
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string"
          },
          "MediaTrackerError": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "MediaTrackerError",
          "errorMessage"
        ]
      },
...
    }
  }
}